feat: improve validation error messaging
Summary:
 When validation errors are encountered the error string that we report
comes straight from pydantic and has two issues:
1. It contains extra information that only serves to make it harder to
   read. e.g. all of the parenthetical text in:
```
steps -> 0 -> parameterSpace -> taskParameterDefinitions -> 0
  Discriminator 'type' is missing in value (type=value_error.discriminated_union.missing_discriminator; discriminator_key=type)
```
2. The path-to-error contains the name of classes for discriminated
   unions. e.g. The 'JobStringParameterDefinition' in:
```
parameterDefinitions -> 0 -> JobStringParameterDefinition -> userInterface -> control
  value is not a valid enumeration member ...
```

 We want a more succint and readable set of error messages.

Solution:
  For (1) we eliminate the paranthetical from the reported error
entirely; pydantic's intent is that that text is for machine processing
of errors, and our customers are not machines. For (2) we introspect the
model's module to determine when a location element is a model class
name and remove it from the location when it is.

 As a bonus, we also condensed the array syntax in the reported location
to use the more commonly known square brackets operator to report the
index.

Signed-off-by: Daniel Neilson <53624638+ddneilson@users.noreply.github.com>