feat: implement semantics for merging a job parameter's definitions (#30)
When creating/running a Job, that Job is defined with a Job Template but
can also include one or more externally-defined Environment Templates.
Both Job Templates and Environment Templates can define Job Parameters
that must be included in the submission, and those Job Parameters have
constraints defined for them in the Templates. If a named job parameter
has definitions in multiple Templates in the submission, then there is
no guarantee that those definitions are identical or even compatible.

When a job parameter has multiple definitions then the Open Job
Description specification has defined semantics for merging those
definitions in to a single definition, and for determining when those
definitions are not compatible with one another.

Basically, the Environment Templates must have an order defined for
them. A Job Parameter's definition gets refined, and more constrained,
as we iterate through this list; culminating in the final refinement by
applying the Job Template's definition (if there is one).

So, this commit adds an internal method that will be used for this job
parameter definition refinement/merger process. It is given a list of
definitions for the same job parameter and merges those definitions in
to a single definition -- checking for compatibility issues as it goes.

Signed-off-by: Daniel Neilson <53624638+ddneilson@users.noreply.github.com>