fix: improve validation errors when parameter defs have errors
Problem:
 We have template validation logic that checks that Format Strings only
reference template variables that have been defined at the location of
the Format String in the model. This implementation is flawed.
 This validation is currently implemented as a post-validation
root_validator on the JobTemplate and EnvironmentTemplate models.
However, post-validators' input is *only* the fields that have *passed*
validation. This means that if, say, we have a validation error in *one*
of the parameter definitions within the template then this validator is
not given *any* of the parameter definitions for the model -- since the
parameter definitions list field fails validation if any list element
fails validtion.
 This manifests in false error reporting since we'll report unknown
reference errors for even template variables whose definition was
correctly formed.

Solution:
 A complete overhaul of the variable reference validation logic.
Pydantic does not provide any way for a post-validator to be given the
pre-validator input, so we need to convert the variable reference
validator to a pre-validator root-validator.

Result:
 A template like:

```yaml
specificationVersion: 'jobtemplate-2023-09'
name: "DemoJob"
parameterDefinitions:
- name: Foo
  type: INT
- name: Bar
  type: INT
  default: "Bleh"
steps:
- name: DemoStep
  script:
    actions:
      onRun:
        command: "echo"
        args:
        - '{{Param.Foo}}'
        - '{{Param.Bar}}'
```

goes from having 3 validation errors:

```
parameterDefinitions[1] -> default:
        value is not a valid integer
__root__ -> steps[0] -> script -> actions -> onRun -> args[0]:
        Param.Foo is referenced by an expression, but is out of scope or has no value
__root__ -> steps[0] -> script -> actions -> onRun -> args[1]:
        Param.Bar is referenced by an expression, but is out of scope or has no value
```

to 1 validation errors (the misleading one is removed:

```
parameterDefinitions[1] -> default:
        value is not a valid integer
```

Signed-off-by: Daniel Neilson <53624638+ddneilson@users.noreply.github.com>