feat!: redefine model versioning enums
Problem:
 The existing `SchemaVersion` enum has become ambiguous. It means both
the Open Job Description revision (e.g. "v2023_09") and the template
specification versions (e.g. "jobtemplate-2023-09"). Originally, the job
template version and the spec revision were synonmous so a single enum
made sense. However, we now have multiple template kinds in a revision
('jobtemplate-2023-09' and 'environment-2023-09'), so things like the
special 'version' field of a model class are now ambiguous.

Solution:
 Split `SchemaVersion` in to two separate enums:
TemplateSpecificationVersion, and RevisionVersion. The former is for the
'specificationVersion' fields of templates, and the later is for the
specific revision of the specification. I've also changed the 'version'
field to 'revision' in all model classes.

**BREAKING CHANGE** - The SchemaVersion enum has been removed. Please
use TemplateSpecificationVersion or SpecificationRevision in its place.

Signed-off-by: Daniel Neilson <53624638+ddneilson@users.noreply.github.com>