fix: update homepage url
Summary:
I'd set the homepage url to the specification repository rather than
this project's repository because I wanted that to be considered the
homepage. However, it looks like PyPI is populating the GitHub
statistics of the project based off of the homepage rather than source
page. The PyPI page's stats should reflect this repo rather than the
other, so let's change the homepage.

Signed-off-by: Daniel Neilson <53624638+ddneilson@users.noreply.github.com>