feat: add validation that associative op's args are equal length
Problem:

The Open Job Description specification says "Every comma-separated expression
within an associative operator must have the exact same number of values defined
in their range." The current model validators to not ensure that this is the case.

Reference: https://github.com/OpenJobDescription/openjd-specifications/wiki/2023-09-Template-Schemas#343-combinationexpr

Solution:

I've implemented validate_step_parameter_space_dimensions() to validate this constraint
and added a validator to the StepParameterSpace to use it. My first inclination was
to implement this validation as part of constructing the StepParameterSpaceIterator
since it already has a recursive expression parse, but that is a chicken & egg problem;
we need a StepParameterSpace to construct a StepParameterSpaceIterator, but we don't
have one yet when doing validation for the construction of a StepParameterSpace.

Signed-off-by: Daniel Neilson <53624638+ddneilson@users.noreply.github.com>