fix: format string errors no longer embed the entire format string
Problem:

Format strings can be very large (e.g. when they are the 'data' element of an embedded file).
The FormatStringError that we report currently embeds the entire format string into the error
message. That's unnecessary since our validation errors are already pointing the user directly
at the template element that is producing the error.

Solution:

This removes the format string data from the error message.

Signed-off-by: Daniel Neilson <53624638+ddneilson@users.noreply.github.com>