docs: Update readme info on creating model objects
Some of the sample code in the README.md no longer worked after the
change for RFC 0002.
https://github.com/OpenJobDescription/openjd-specifications/blob/mainline/rfcs/0002-model-extensions.md
updated OpenJD to make the fields and validation of template components
depend on an 'extensions' property at the top level of the template.
Implementing this in the model library involved adding a model parsing
context that objects look at during construction and validation.

This change adds a docstring for the parse_model function and updates
the broken examples to explain how the extensions fit in and how to
create model objects.

Signed-off-by: Mark <399551+mwiebe@users.noreply.github.com>